+++ lib/src/mobile_scanner_controller.dart
@@ -442,9 +442,10 @@ class MobileScannerController {
                   rawBytes: barcode['rawBytes'] as Uint8List?,
                   format: toFormat(barcode['format'] as int),
                   corners: toCorners(
-                    (barcode['corners'] as List<Object?>? ?? [])
-                        .cast<Map<Object?, Object?>>(),
-                  ),
+                        (barcode['corners'] as List<Object?>? ?? [])
+                            .cast<Map<Object?, Object?>>(),
+                      ) ??
+                      const <Offset>[],
                 ),
             ],
           ),